+++ .gitignore
@@ -3,7 +3,6 @@ div-tracker-env
 __pycache__/
 .env
 db.sqlite3
-/static/
 /media/
 # IDE-specific files
 .vscode/
