+++ portfolio/templates/portfolio/layout.html
@@ -9,41 +9,46 @@
     <title>Portfolio</title>
 </head>
 <body>
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <div id="navBarLogo">
-            <a class="navbar-brand" href="{% url 'portfolio:index' %}">
-                <img src="{% static 'portfolio/img/dollar-symbol.png' %}" height="30px" alt="Your Logo">
-            </a>
-        </div>
-
-        <!-- Navbar links -->
-        <div class="collapse navbar-collapse" id="navbarNav">
-            <ul class="navbar-nav ml-auto">
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'portfolio:load_buy_stock' %}">Buy Stock</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'portfolio:load_sell_stock' %}">Sell Stocks</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'portfolio:view_dividends' %}">View Dividends</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'portfolio:view_dividends' %}">View Metrics</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'users:logout' %}">Logout</a>
-                </li>
-            </ul>
-        </div>
-
-        <!-- Mobile menu button div -->
-        <div class="mobile-menu-button">
-            <button class="navbar-toggler" type="button" aria-expanded="false" aria-label="Toggle navigation">
-                <img src="{% static 'portfolio/img/mobile-menu.png' %}" alt="Mobile Menu Icon" id="mobileMenuImg">
-            </button>
-        </div>
-    </nav>
+    <header class="header">
+        <nav class="navbar navbar-expand-lg navbar-light bg-light">
+            <div id="navBarLogo">
+                <a class="navbar-brand" href="{% url 'portfolio:index' %}">
+                    <img src="{% static 'portfolio/img/dollar-symbol.png' %}" height="30px" alt="Your Logo">
+                </a>
+            </div>
+    
+            <!-- Navbar links -->
+            <div class="collapse navbar-collapse" id="navbarNav">
+                <ul class="nav-menu ml-auto">
+                    <li class="nav-item">
+                        <a class="nav-link" href="{% url 'portfolio:load_buy_stock' %}">Buy Stock</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="{% url 'portfolio:load_sell_stock' %}">Sell Stocks</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="{% url 'portfolio:view_dividends' %}">View Dividends</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="{% url 'portfolio:view_dividends' %}">View Metrics</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="{% url 'users:logout' %}">Logout</a>
+                    </li>
+                </ul>
+            </div>
+    
+            <!-- Mobile menu button div -->
+            <div class="mobile-menu-button">
+                <span class="mobile-menu-bar"></span>
+                <span class="mobile-menu-bar"></span>
+                <span class="mobile-menu-bar"></span>
+                <!-- <img src="{% static 'portfolio/img/mobile-menu.png' %}" alt="Mobile Menu Icon" id="mobileMenuImg"> -->
+                <!-- <button class="navbar-toggler" type="button" aria-expanded="false" aria-label="Toggle navigation">
+                </button> -->
+            </div>
+        </nav>
+    </header>
     <div class="container mt-5">
         {% block body %}
         {% endblock %}
@@ -53,53 +58,27 @@
 <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
 <!-- JavaScript/jQuery code -->
 <script>
-    $(document).ready(function() {
-    // Toggle the navbar links when the mobile menu button is clicked
-    $(".navbar-toggler").click(function() {
-        // Toggle the display of the navbar links
-        $("#navbarNav").toggleClass("show");
+    const mobileMenuDiv = document.querySelector(".mobile-menu-button");
+    const navMenu = document.querySelector(".nav-menu");
+    const navMenuColap = document.querySelector(".navbar-collapse");
 
-        // If the navbar links are visible, adjust the layout
-        if ($("#navbarNav").hasClass("show")) {
-            // Calculate the position to display the navbar links
-            var logoHeight = $("#navBarLogo").outerHeight();
-            var navbarHeight = $("#navbarNav").outerHeight();
-            var totalHeight = logoHeight + navbarHeight;
+    mobileMenuDiv.addEventListener("click", mobileMenu);
 
-            // Adjust the position of the navbar links
-            $("#navbarNav").css({
-                position: "absolute",
-                top: totalHeight + "px",
-                left: "0",
-                zIndex: "1000",
-                backgroundColor: "#fff",
-                width: "100%",
-                display: "none"
-            }).slideDown(); // Slide down to show the navbar links
-        } else {
-            // Hide the navbar links and reset the layout
-            $("#navbarNav").slideUp(function() {
-                $(this).css({
-                    position: "",
-                    top: "",
-                    left: "",
-                    zIndex: "",
-                    backgroundColor: "",
-                    width: "",
-                    display: ""
-                });
-            });
-        }
-    });
+    function mobileMenu() {
+        console.log("Enter mobileMenu function");
+        mobileMenuDiv.classList.toggle("active");
+        navMenu.classList.toggle("active");
+        navMenuColap.classList.toggle("active");
+    }
 
-    // Hide the navbar links when the window is resized to wider than 992px
-    $(window).resize(function() {
-        if ($(window).width() > 992) {
-            $("#navbarNav").removeClass("show").removeAttr("style");
-        }
-    });
-});
+    const navLink = document.querySelectorAll(".nav-link");
 
+    navLink.forEach(n => n.addEventListener("click", closeMenu));
+
+    function closeMenu() {
+        mobileMenuDiv.classList.remove("active");
+        navMenuColap.classList.remove("active");
+    }
 </script>
 </body>
 </html>
