+++ static/portfolio/css/navbar_style.css
@@ -1,0 +1,98 @@
+/* Custom styles for the navbar */
+.header {
+    border-bottom: 1px solid #E2E8F0;
+}
+
+.navbar {
+    height: 50px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.navbar-brand {
+    margin-right: auto; /* Moves the logo to the right */
+}
+
+.mobile-menu-button {
+    display: none;
+}
+
+.mobile-menu-bar {
+    display: block;
+    width: 25px;
+    height: 3px;
+    margin: 5px auto;
+    -webkit-transition: all 0.3s ease-in-out;
+    transition: all 0.3s ease-in-out;
+    background-color: #101010;
+}
+
+.nav-menu {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.nav-item {
+    list-style-type: none; /* Removes the dots */
+    margin-right: 15px; /* Adds some space between the items */
+}
+
+.nav-link{
+    /* text-decoration: none; */
+    padding: 15px 20px; /* Adjust padding for better aesthetic */
+    color: #333; /* Text color */
+    text-decoration: none; /* Removes underline */
+    /* font-size: 1.6rem;
+    font-weight: 400;
+    color: #475569; */
+}
+
+.nav-menu .nav-item:hover {
+    color: #242424;
+    transform: scale(1.2); /* Zoom effect on hover */
+}
+
+@media only screen and (max-width: 768px) {
+    .nav-menu {
+        position: fixed;
+        left: -110%;
+        top: 62px;
+        flex-direction: column;
+        background-color: #fff;
+        width: 100%;
+        border-radius: 10px;
+        text-align: center;
+        transition: 0.3s;
+        box-shadow:
+            0 10px 27px rgba(0, 0, 0, 0.05);
+    }
+
+    .nav-menu.active {
+        left: 0;
+        padding: 0;
+    }
+
+    .nav-item {
+        margin: 10px 0;
+    }
+
+    .mobile-menu-button {
+        display: block;
+        cursor: pointer;
+    }
+
+    .mobile-menu-button.active .mobile-menu-bar:nth-child(2) {
+        opacity: 0;
+    }
+
+    .mobile-menu-button.active .mobile-menu-bar:nth-child(1) {
+        transform: translateY(8px) rotate(45deg);
+    }
+
+    .mobile-menu-button.active .mobile-menu-bar:nth-child(3) {
+        transform: translateY(-8px) rotate(-45deg);
+    }
+
+}
